+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicStateStatus.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.engine.management.periodic
 
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
 import pl.touk.nussknacker.engine.api.deployment.StateStatus.StatusName
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
 import pl.touk.nussknacker.engine.api.deployment.{CustomStateStatus, FailedStateStatus, ProcessActionType, RunningStateStatus, StateDefinitionDetails, StateStatus}
 
 import java.net.URI
@@ -21,7 +22,8 @@ object PeriodicStateStatus {
     case _: RunningStateStatus => List(ProcessActionType.Cancel) //periodic processes cannot be redeployed from GUI
     case _: ScheduledStatus => List(ProcessActionType.Cancel, ProcessActionType.Deploy)
     case WaitingForScheduleStatus => List(ProcessActionType.Cancel) //or maybe should it be empty??
-    case _: FailedStateStatus => List(ProcessActionType.Cancel) //original FailedStateStatus allows redeployment
+    case _: FailedStateStatus => List(ProcessActionType.Cancel) //redeploy is not allowed
+    case _: ProblemStateStatus => List(ProcessActionType.Cancel) //redeploy is not allowed
   }
 
   val statusTooltipsPF: PartialFunction[StateStatus, Option[String]] = {