+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkStateStatus.scala
@@ -1,15 +1,13 @@
 package pl.touk.nussknacker.engine.management
 
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
-import pl.touk.nussknacker.engine.api.deployment.StateStatus.StatusName
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 
-import java.net.URI
-
 object FlinkStateStatus {
-  val FailedToGet: StateStatus = SimpleStateStatus.FailedToGet
-  val Unknown: StateStatus = SimpleStateStatus.Unknown
+
+  // Flink statuses are based on SimpleStateStatus definitions
+  // TODO: remove those states here and use SimpleStateStatus
   val NotDeployed: StateStatus = SimpleStateStatus.NotDeployed
   val DuringDeploy: StateStatus = SimpleStateStatus.DuringDeploy
   val Running: StateStatus = SimpleStateStatus.Running
@@ -17,31 +15,10 @@ object FlinkStateStatus {
   val Restarting: StateStatus = SimpleStateStatus.Restarting
   val DuringCancel: StateStatus = SimpleStateStatus.DuringCancel
   val Canceled: StateStatus = SimpleStateStatus.Canceled
-  val Failed: StateStatus = SimpleStateStatus.Failed
-  val Error: StateStatus = SimpleStateStatus.Error
-  val Warning: StateStatus = SimpleStateStatus.Warning
-
-  val Failing: StateStatus =  NotEstablishedStateStatus("FAILING")
-  val MultipleJobsRunning: StateStatus = NotEstablishedStateStatus("MULTIPLE_JOBS_RUNNING")
 
   val statusActionsPF: PartialFunction[StateStatus, List[ProcessActionType]] = {
-    case SimpleStateStatus.DuringDeploy => List(ProcessActionType.Cancel)
-    case SimpleStateStatus.Restarting => List(ProcessActionType.Cancel)
-    case FlinkStateStatus.MultipleJobsRunning => List(ProcessActionType.Cancel)
+    case FlinkStateStatus.DuringDeploy => List(ProcessActionType.Cancel)
+    case FlinkStateStatus.Restarting => List(ProcessActionType.Cancel)
   }
 
-  val customStateDefinitions: Map[StatusName, StateDefinitionDetails] = Map(
-    Failing.name -> StateDefinitionDetails(
-      displayableName = "Failing",
-      icon = Some(URI.create("/assets/states/error.svg")),
-      tooltip = Some("Failing"),
-      description = Some("Failing")
-    ),
-    MultipleJobsRunning.name -> StateDefinitionDetails(
-      displayableName = "More than one deployment running",
-      icon = Some(URI.create("/assets/states/error.svg")),
-      tooltip = Some("More than one deployment running"),
-      description = Some("More than one deployment running")
-    )
-  )
 }